# docs/swagger.yaml
definitions:
  handlers.AccessRequestPayload:
    properties:
      canSelfApprove:
        type: boolean
      cidr:
        type: string
      description:
        type: string
      direction:
        type: string
      duration:
        type: integer
      ports:
        type: string
      requestID:
        type: string
      requestType:
        type: string
      requestor:
        type: string
      service:
        type: string
      sourceService:
        type: string
      status:
        type: string
      targetService:
        type: string
      timestamp:
        type: integer
    type: object
  handlers.ActiveAccessInfo:
    properties:
      direction:
        type: string
      expiresAt:
        type: integer
      name:
        type: string
      namespace:
        type: string
      ports:
        type: string
      source:
        type: string
      status:
        type: string
      target:
        type: string
      type:
        type: string
    type: object
  handlers.HTTPError:
    properties:
      error:
        example: Error message
        type: string
    type: object
  handlers.LogEntry:
    properties:
      className:
        type: string
      logType:
        type: string
      payload:
        type: string
      timestamp:
        type: integer
      type:
        type: string
    type: object
  handlers.ServiceInfo:
    properties:
      compound:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
    type: object
info:
  contact: {}
paths:
  /active-accesses:
    get:
      description: Retrieves all active and partially-created (pending) access policies
        managed by Netwatch.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.ActiveAccessInfo'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: List active access policies
      tags:
      - Access Policies
  /logs:
    get:
      description: Retrieves all persisted log entries from the application.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.LogEntry'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Get global activity log
      tags:
      - System
  /pending-requests:
    get:
      description: Retrieves all pending AccessRequest custom resources and enriches
        them with the current user's permissions.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.AccessRequestPayload'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: List pending access requests
      tags:
      - Requests
  /services:
    get:
      description: Retrieves a list of all services in the cluster, filtered to exclude
        system and clone services.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.ServiceInfo'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: List all Kubernetes services
      tags:
      - System
swagger: "2.0"